Restructure email CLI into namespace with test, preview, config commands
Move email commands from flat files into CLI::Email module with
lib/onetime/cli/email/ subdirectory. Extract shared AVAILABLE_TEMPLATES
and SAMPLES_PATH into email.rb module.

New commands:
- `ots email test` — SMTP connectivity check with timing
- `ots email preview` — render templates with sample data or --data JSON,
  supports --html and --open for browser preview
- `ots email config` — display delivery provider and masked credentials

Add YAML sample data files for all 11 templates in lib/onetime/mail/samples/
so `email preview` works without requiring --data.
