+++ lib/onetime/mail/samples/email_change_requested.yml
@@ -1,0 +1,3 @@
+old_email: "current@example.com"
+new_email: "newemail@example.com"
+requested_at: "2026-01-15T10:30:00Z"
