+++ lib/onetime/mail/samples/email_changed.yml
@@ -1,0 +1,3 @@
+old_email: "previous@example.com"
+new_email: "current@example.com"
+changed_at: "2026-01-15T10:45:00Z"
