+++ lib/onetime/mail/samples/expiration_warning.yml
@@ -1,0 +1,3 @@
+recipient: "owner@example.com"
+secret_key: "expiring-key-sample"
+expires_at: 1800000000
