+++ lib/onetime/mail/samples/incoming_secret.yml
@@ -1,0 +1,3 @@
+recipient: "recipient@example.com"
+sender_email: "sender@example.com"
+secret_key: "incoming-key-sample"
