+++ lib/onetime/mail/samples/password_request.yml
@@ -1,0 +1,2 @@
+email_address: "user@example.com"
+reset_password_path: "https://example.com/reset-password/sample-token-xyz789"
