+++ lib/onetime/mail/samples/secret_link.yml
@@ -1,0 +1,4 @@
+secret_key: "abc123def456"
+recipient: "recipient@example.com"
+sender_email: "sender@example.com"
+share_domain: ""
