+++ lib/onetime/mail/samples/secret_revealed.yml
@@ -1,0 +1,3 @@
+recipient: "owner@example.com"
+secret_shortid: "abc123"
+revealed_at: "2026-01-15T14:30:00Z"
